Metodos atacar no Survivor e no Infected
new: FALTA POR RANDOM NOS METODOS DE ATACAR DO SURVIVOR PRA INCLUIR A PRECISÃO NO ATAQUE.

FALTA COLOCAR REMOVER ITENS QUANDO USAR ALGUM ITEM, FALTA FAZER AS INTERAÇÕES DE
GUNS COM AMMUNITION, FALTA MEXER NOS AMBIENTES, FALTA COLOCAR RANDOM NA
DECLARAÇÃO DE MONSTROS NOS AMBIENTES, FALTA CRIAR UMA INTERFACE MENU
PARA O PERSONAGEM.

+++ JavaApplication2/src/Item/ItemAntidote.java
@@ -33,9 +33,9 @@ public class ItemAntidote extends Item {
 
     @Override
     public int usar(Item item, Survivor personagem) {
-        int x = personagem.getInfection();
+        double x = personagem.getInfection();
         int y = item.getIntensity();
-        int z = x-y;
+        int z = (int) (x-y);
         if(z<0){
             z = 0;
             personagem.setInfection(z);
